backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
asyncpg==0.29.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pytest==7.4.3
httpx==0.25.2
PyJWT==2.8.0
python-dotenv==1.0.0
black==23.10.1
flake8==6.0.0
isort==5.12.0